# core-module/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Национальный сервис мониторинга цен на продовольственные товары(OpenAPI 3.0)
  description: |-
    Национальный сервис мониторинга цен на продовольственные товары
    
    Полезные ссылки:
    - [Репозиторий](https://github.com/AnnaDesert/Learning)
  version: 1.0.11
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /product:
    get:
      tags:
        - product
      summary: Получить все товары
      description: Получить все товары
      operationId: allProducts
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
    post:
      tags:
        - product
      summary: Добваить продукт
      description: Добваить новый продукт
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreationDTO'
        required: true
      responses:
        '201':
          description: Created
  /product/{id}:
    get:
      tags:
        - product
      summary: Получить товар по ID
      description: Получить товар по ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
    delete:
      tags:
        - product
      summary: Удалить товар по ID
      description: Удалить товар по ID
      operationId: removeProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - product
      summary: Изменить продукт
      description: Изменить продукт
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreationDTO'
        required: true
      responses:
        '200':
          description: OK
  /product/name/{name}:
    get:
      tags:
        - product
      summary: Получить товар по названию
      description: Получить товар по названию
      operationId: getProductByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
  /product/category/{category}:
    get:
      tags:
        - product
      summary: Получить товар по категории
      description: Получить товар по категории
      operationId: getProductByCategory
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
  /product/{fileName}:
    post:
      tags:
        - product
      summary: Добавить товары из csv файла
      description: Добавить товары из csv файла
      operationId: addProductFromCSV
      parameters:
        - name: fileName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /batch:
    get:
      tags:
        - batch
      summary: Получить все партии
      description: Получить все партии
      operationId: allBatch
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchOfProductDTO'
    post:
      tags:
        - batch
      summary: Добваить партию
      description: Добваить партию
      operationId: addBatch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchOfProductCreationDTO'
        required: true
      responses:
        '201':
          description: Created
  /batch/{id}:
    get:
      tags:
        - batch
      summary: Получить партию по ID
      description: Получить партию по ID
      operationId: getBatchById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/BatchOfProductDTO'
    delete:
      tags:
        - batch
      summary: Удалить партию по ID
      description: Удалить партию по ID
      operationId: removeBatchById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      responses:
        '200':
          description: OK
    patch:
      tags:
        - batch
      summary: Изменить партию
      description: Изменить партию
      operationId: updateBatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            minimum: 1
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchOfProductCreationDTO'
        required: true
      responses:
        '200':
          description: OK
  /batch/productCategory/{productCategory}:
    get:
      tags:
        - batch
      summary: Получить партии по категории
      description: Получить партии по категории
      operationId: getBatchByProductCategory
      parameters:
        - name: productCategory
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BatchOfProductDTO'
...
components:
  schemas:
...
    ProductCreationDTO:
      required:
        - name
      type: object
      properties:
        name:
          type: string
...
    BatchOfProductCreationDTO:
      required:
        - count
        - expirationDateEnd
        - expirationDateStart
        - idProduct
      type: object
      properties:
        expirationDateStart:
          type: string
          format: date-time
        expirationDateEnd:
          type: string
          format: date-time
        idProduct:
          type: integer
          format: int64
        count:
          type: integer
          format: int32
...
    ProductDTO:
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    BatchOfProductDTO:
      required:
        - count
        - expirationDateEnd
        - expirationDateStart
        - id
        - idProduct
      type: object
      properties:
        id:
          type: integer
          format: int64
        expirationDateStart:
          type: string
          format: date-time
        expirationDateEnd:
          type: string
          format: date-time
        idProduct:
          type: integer
          format: int64
        count:
          type: integer
          format: int32